requirements.txt
# Core numeric / ML stack
numpy
pandas
scikit-learn
torch
faiss-cpu

# NLP + embeddings
sentence-transformers
spacy

# Evaluation & plotting
matplotlib
seaborn
matplotlib-venn
tqdm

# Tooling & config
rich
python-dotenv
PyYAML
